Saved lists (#36)
* create saved list

* move favorite front script to different file
