+++ routes/pages/lists.routes.js
@@ -1,18 +1,17 @@
 const express = require("express");
 const router = express.Router();
 const List = require("../../models/List.model");
-const User = require("../../models/User.model");
 const Ingredient = require("../../models/Ingredient.model");
 const Category = require("../../models/Category.model");
 const Unit = require("../../models/Unit.model");
-const { isLoggedIn } = require("../middlewares/auth");
 
-// Route prefix : /lists
+// Route prefix: /lists
 
 router.get("/", async (req, res, next) => {
+    // display all lists
     try {
-        // display all lists
         res.locals.navbar.link = "/lists/create";
+        res.locals.scripts = ["/js/favorites.js"];
 
         const user = req.session.currentUser.id;
         const userLists = await List.find({
@@ -28,9 +27,10 @@ router.get("/create", (req, res, next) => {
     try {
         res.locals.navbar.link = `/lists`;
         res.locals.navbar.icon = "fa-solid fa-check";
-        res.locals.scripts = ["/js/list-create.js"];
+        res.locals.scripts = ["/js/list-create.js", "/js/favorites.js"];
         res.locals.list = {
             name: "New List",
+            template: false,
         };
 
         res.render("lists/list-edit");
@@ -44,7 +44,7 @@ router.get("/:listId", async (req, res, next) => {
     try {
         res.locals.navbar.icon = "fa-regular fa-pen-to-square";
         res.locals.navbar.link = `/lists/edit/${req.params.listId}`;
-        res.locals.scripts = ["/js/list-details.js"];
+        res.locals.scripts = ["/js/list-details.js", "/js/favorites.js"];
 
         // #TODO aggregate with mongoose
         const list = await List.findById(req.params.listId)
@@ -68,7 +68,7 @@ router.get("/edit/:listId", async (req, res, next) => {
     try {
         res.locals.navbar.icon = "fa-solid fa-check";
         res.locals.navbar.link = `/lists/${req.params.listId}`;
-        res.locals.scripts = ["/js/list-edit.js"];
+        res.locals.scripts = ["/js/list-edit.js", "/js/favorites.js"];
 
         // #TODO aggregate with mongoose
         const list = await List.findById(req.params.listId)
@@ -86,11 +86,6 @@ router.get("/edit/:listId", async (req, res, next) => {
     } catch (error) {
         next(error);
     }
-});
-
-router.patch("/edit/:listId", (req, res, next) => {
-    // list edition form submission
-    res.send(req.body);
 });
 
 module.exports = router;
